Fix dead foam rendering by rescaling foam thresholds
The foam presets {30, 5, 4, 3, 2} were sized for PARTICLE_THRESHOLD 5.0f
and integer LED particle counts. d96cc93 dropped the threshold to 2.0f and
5af7ced swapped the counts for float intensities, leaving every preset
above the liquid threshold so the foam branch never ran.

Rescale the presets below PARTICLE_THRESHOLD as floats, keeping index 0
above it as the no foam option. Indices are unchanged so saved NVS
settings still load.

+++ include/button.h
@@ -73,7 +73,7 @@ class ButtonHandler {
   
   // preset arrays
   int brightness_presets[7];
-  int foam_presets[5];
+  float foam_presets[5];
   
   // nvs storage object
   Preferences preferences;
@@ -102,7 +102,7 @@ class ButtonHandler {
   // getters for current settings
   ColorPreset getCurrentColor();
   int getCurrentBrightness();
-  int getCurrentFoamThreshold();
+  float getCurrentFoamThreshold();
   
   // get rgb values for current color preset
   void getCurrentColorRGB(uint8_t* r, uint8_t* g, uint8_t* b);
